fix: mark unsigned

+++ include/hash_table.h
@@ -24,8 +24,8 @@ ht_hash_table *ht_new();
 static ht_item *ht_new_item(const uint8_t *, const uint8_t *);
 static void ht_del_item(ht_item *);
 void ht_del_hash_table(ht_hash_table *);
-static int32_t ht_hash(const uint8_t *, const int32_t, const int32_t);
-static int32_t ht_get_hash(const uint8_t *, const size_t, const int32_t);
+static size_t ht_hash(const uint8_t *, const int32_t, const int32_t);
+static size_t ht_get_hash(const uint8_t *, const size_t, const int32_t);
 void ht_insert(ht_hash_table *, const uint8_t *, const uint8_t *);
 
 #endif
